Replace `test_hub` with equivalent `test_scope`
